chords_generator/pyproject.toml
[project]
name = "guitar-player-chords-generator"
version = "0.1.0"
description = "Chord recognition service using autochord"
requires-python = ">=3.10"
dependencies = [
    "setuptools<81",
    "numpy>=1.24.0",
    "tf-keras>=2.16.0",
    "autochord>=0.1.4",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "boto3>=1.35.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "mangum>=0.19.0",
    "pychord>=1.3.2",
    "python-json-logger>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "moto[s3]>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chords_generator"]

# vamp's setup.py imports numpy/setuptools at build time without declaring
# them as build deps. Disable build isolation so it uses the environment's.
[tool.uv]
no-build-isolation-package = ["vamp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11"
extraPaths = ["src"]
